prune-trace: use uint64_t for pilot counts
This enables using very large data sets in the FESamplingPruner.

+++ tools/prune-trace/FESamplingPruner.cc
@@ -93,7 +93,7 @@ bool FESamplingPruner::sampling_prune(const fail::Database::Variant& variant)
 	MYSQL_RES *res;
 	MYSQL_ROW row;
 
-	unsigned pilotcount = 0, samplerows;
+	uint64_t pilotcount = 0, samplerows;
 
 	if (!m_use_known_results) {
 		LOG << "loading trace entries for " << variant.variant << "/" << variant.benchmark << " ..." << endl;
@@ -149,7 +149,7 @@ bool FESamplingPruner::sampling_prune(const fail::Database::Variant& variant)
 	LOG << "loaded " << pilotcount << " entries, sampling "
 		<< samplerows << " entries with fault expansion ..." << endl;
 
-	unsigned num_fspgroup_entries = 0;
+	uint64_t num_fspgroup_entries = 0;
 	uint32_t known_pilot_method_id = m_method_id;
 
 	if (!m_use_known_results) {
@@ -159,7 +159,7 @@ bool FESamplingPruner::sampling_prune(const fail::Database::Variant& variant)
 		std::string insert_sql(ss.str());
 		ss.str("");
 
-		for (unsigned i = 0; i < samplerows; ++i) {
+		for (uint64_t i = 0; i < samplerows; ++i) {
 			uint64_t pos = my_rand(pop.get_size() - 1);
 			Pilot p = pop.get(pos);
 			ss << "(0," << variant.id << "," << p.instr2 << "," << p.instr2
@@ -170,7 +170,7 @@ bool FESamplingPruner::sampling_prune(const fail::Database::Variant& variant)
 		}
 		db->insert_multiple();
 
-		unsigned num_fsppilot_entries = samplerows;
+		uint64_t num_fsppilot_entries = samplerows;
 
 		// single entry for known outcome (write access)
 		ss << "INSERT INTO fsppilot (known_outcome, variant_id, instr2, injection_instr, injection_instr_absolute, data_address, data_width, fspmethod_id) "
@@ -201,7 +201,7 @@ bool FESamplingPruner::sampling_prune(const fail::Database::Variant& variant)
 		std::string insert_sql(ss.str());
 		ss.str("");
 
-		for (unsigned i = 0; i < samplerows; ++i) {
+		for (uint64_t i = 0; i < samplerows; ++i) {
 			uint64_t pos = my_rand(pop.get_size() - 1);
 			Pilot p = pop.get(pos);
 			ss << "(" << variant.id << "," << p.instr2
